pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "octmultitask"
version = "0.1.0"
description = "Multitask OCT segmentation & classification"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "numpy>=1.24",
  "opencv-python>=4.8",
  "albumentations>=1.4",
  "pillow>=10.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
optuna = ["optuna>=3.6"]
api = ["pydantic>=2.7"]

[tool.setuptools]
packages = ["datasets", "models", "training", "inference", "evaluation", "service", "utils", "data"]
